use daemon threads to allow shut down of application (#748)
To exit normally all non-daemon threads of the java process are supposed
to have stopped. TcpClientSession creates an static class member
EventLoopGroup which in turn instantiates worker threads. These are
changed to daemon threads to not get in the way of the shutdown. This
is similar to the default in Netty 5.

To ensure work is finished the Runtime shutdown event is hooked to
initiate a time-limited graceful shutdown of the event loops. Normally
an application should shut down communication before the process
shutdown is inititated.

Co-authored-by: Peter Svensson <petersv@psv.nu>
Co-authored-by: Konicai <71294714+konicai@users.noreply.github.com>

(cherry picked from commit d85e90f3412fe26a055dceabfe1a7a4793c87651)